+++ healthcheck/wrappers.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from .healthcheck import Checker
+
+
+def checker(name=None):
+    if callable(name):
+        return Checker().decorate(name)
+    return Checker(name=name).decorate
